Minor fixes, refactored some of the launchers. Round the item prices. Added a separate composite for the merged store.
git-svn-id: https://svn.apache.org/repos/asf/incubator/tuscany/java/sca@589202 13f79535-47bb-0310-9956-ffa450edef68

+++ tutorial/assets/services/ShoppingCartImpl.java
@@ -84,7 +84,7 @@ public class ShoppingCartImpl implements Collection {
     }
 
     private String getTotal() {
-        float total = 0;
+        double total = 0;
         String currencySymbol = "";
         if (!cart.isEmpty()) {
             String item = ((Content)cart.values().iterator().next().getContents().get(0)).getValue();
@@ -92,7 +92,7 @@ public class ShoppingCartImpl implements Collection {
         }
         for (Entry entry : cart.values()) {
             String item = ((Content)entry.getContents().get(0)).getValue();
-            total += Float.valueOf(item.substring(item.indexOf("-") + 3));
+            total += Double.valueOf(item.substring(item.indexOf("-") + 3));
         }
         return currencySymbol + String.valueOf(total);
     }
